chore(chapter10): rename case03.py to case03_1.py and update import in test_case03.py
